pyproject.toml
[project]
name = "backend-drone-vla-3d"
version = "0.1.0"
description = "Backend for simulator web 3D drone VLA (Visual Language Action)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "numpy>=1.26.4,<2",
    "pillow>=11.0.0",
    "modal>=0.73.0",
]
